Added some more changes and bug fixes

+++ database/migrations/2024_02_19_150603_create_address_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email');
-            $table->integer('phone');
+            $table->decimal('phone', 10, 2);
             $table->string('address');
             $table->string('state');
             $table->integer('pincode');
